Updating firewx plot script for restart

+++ scripts/cam/plots/exevs_cam_nam_firewxnest_grid2obs_plots.sh
@@ -47,6 +47,7 @@ done
 
 for varb in TMP DPT RH
 do
+	mkdir -p $COMOUTplots/$varb
 	export var=${varb}2m
 	export lev=Z2
 	export lev_obs=Z2
@@ -55,42 +56,102 @@ do
 	smvar=`echo $varb | tr A-Z a-z`
 	export plottyp=lead
 	export datetyp=VALID
-	sh $USHevs/${COMPONENT}/py_plotting.config
 
+	if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png ]
+	then
+	sh $USHevs/${COMPONENT}/py_plotting.config
+        else
+	echo "RESTART - plot exists; copying over to plot directory"
+	cp $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png $PLOTDIR
+	fi
+
+	if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+	then
         mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png
+	cp ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png $COMOUTplots/$varb
+        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png ]
+	then
+	echo "NO PLOT FOR",$var,$plottyp
+	fi
 
 	export plottyp=valid_hour
 	export datetyp=INIT
-	sh $USHevs/${COMPONENT}/py_plotting.config
 
+	if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png ]
+	then
+	sh $USHevs/${COMPONENT}/py_plotting.config
+        else
+	echo "RESTART - plot exists; copying over to plot directory"
+	cp $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png $PLOTDIR
+	fi
+
+	if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+	then
         mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png
+	cp ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png $COMOUTplots/$varb
+        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png ]
+	then
+	echo "NO PLOT FOR",$var,$plottyp
+	fi
 
 	if [ $varb = DPT ]
 	then 
+		mkdir -p $COMOUTplots/$varb
 		export plottyp=threshold_average
 		export datetyp=INIT
 		export linetype=CTC
 		export stat=fbias
                 export thresh=">=277.594, >=283.15, >=288.706, >=294.261"
-		sh $USHevs/${COMPONENT}/py_plotting.config_thresh
 
+		if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png ]
+		then
+		sh $USHevs/${COMPONENT}/py_plotting.config_thresh
+	        else
+		echo "RESTART - plot exists; copying over to plot directory"
+		cp $COMOUTplots/$varb/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png $PLOTDIR
+		fi
+
+		if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+		then
 		mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png
+		cp ${PLOTDIR}/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png $COMOUTplots/$varb
+	        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png ]
+		then
+		echo "NO PLOT FOR",$var,$plottyp
+		fi
+
 	elif [ $varb = RH ]
 	then
+		mkdir -p $COMOUTplots/$varb
 		export plottyp=threshold_average
 		export datetyp=INIT
 		export linetype=CTC
 		export stat=fbias
 		export thresh="<=15, <=20, <=25, <=30"
-		sh $USHevs/${COMPONENT}/py_plotting.config_thresh
 
+		if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png ]
+		then
+		sh $USHevs/${COMPONENT}/py_plotting.config_thresh
+	        else
+		echo "RESTART - plot exists; copying over to plot directory"
+		cp $COMOUTplots/$varb/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png $PLOTDIR
+		fi
+
+		if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+		then
 		mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png
+		cp ${PLOTDIR}/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png $COMOUTplots/$varb
+	        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.fbias.${smvar}_${smlev}.last31days.threshmean.firewx.png ]
+		then
+		echo "NO PLOT FOR",$var,$plottyp
+		fi
 	fi
 done
 
 
 for varb in UGRD VGRD UGRD_VGRD WIND
 do
+	mkdir -p $COMOUTplots/$varb
 	export var=${varb}10m
 	export lev=Z10
 	export lev_obs=Z10
@@ -104,20 +165,49 @@ do
 	smvar=`echo $varb | tr A-Z a-z`
 	export plottyp=lead
 	export datetyp=VALID
-	sh $USHevs/${COMPONENT}/py_plotting.config
 
+	if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png ]
+	then
+	sh $USHevs/${COMPONENT}/py_plotting.config
+        else
+	echo "RESTART - plot exists; copying over to plot directory"
+	cp $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png $PLOTDIR
+	fi
+
+        if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+	then
         mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png
+	cp ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png $COMOUTplots/$varb
+        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png ]
+	then
+	echo "NO PLOT FOR",$var,$plottyp
+	fi
 
 	export plottyp=valid_hour
 	export datetyp=INIT
-	sh $USHevs/${COMPONENT}/py_plotting.config
 
+	if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png ]
+	then
+	sh $USHevs/${COMPONENT}/py_plotting.config
+        else
+	echo "RESTART - plot exists; copying over to plot directory"
+	cp $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png $PLOTDIR
+	fi
+
+        if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+	then
 	mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png
+	cp ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png $COMOUTplots/$varb
+        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png ]
+	then
+	echo "NO PLOT FOR",$var,$plottyp
+	fi
 
 done
 
 for varb in GUST
 do
+	mkdir -p $COMOUTplots/$varb
 	export var=${varb}sfc
 	export lev=Z0
 	export lev_obs=Z0
@@ -126,19 +216,48 @@ do
 	smvar=`echo $varb | tr A-Z a-z`
 	export plottyp=lead
 	export datetyp=VALID
-	sh $USHevs/${COMPONENT}/py_plotting.config
 
+	if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png ]
+	then
+	sh $USHevs/${COMPONENT}/py_plotting.config
+        else
+	echo "RESTART - plot exists; copying over to plot directory"
+	cp $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png $PLOTDIR
+	fi
+
+	if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+	then
         mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png
+	cp ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png $COMOUTplots/$varb
+        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png ]
+	then
+	echo "NO PLOT FOR",$var,$plottyp
+	fi
 
 	export plottyp=valid_hour
 	export datetyp=INIT
-	sh $USHevs/${COMPONENT}/py_plotting.config
 
+	if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png ]
+	then
+	sh $USHevs/${COMPONENT}/py_plotting.config
+        else
+	echo "RESTART - plot exists; copying over to plot directory"
+	cp $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png $PLOTDIR
+	fi
+
+        if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+	then
 	mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png
+	cp ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png $COMOUTplots/$varb
+        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png ]
+	then
+	echo "NO PLOT FOR",$var,$plottyp
+	fi
 done
 
 for varb in PBL
 do
+	mkdir -p $COMOUTplots/$varb
 	export var=${varb}
 	export lev=L0
 	export lev_obs=L0
@@ -147,15 +266,43 @@ do
 	smvar=`echo $varb | tr A-Z a-z`
 	export plottyp=lead
 	export datetyp=VALID
-	sh $USHevs/${COMPONENT}/py_plotting.config_pbl
 
+	if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png ]
+	then
+	sh $USHevs/${COMPONENT}/py_plotting.config_pbl
+        else
+	echo "RESTART - plot exists; copying over to plot directory"
+	cp $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png $PLOTDIR
+	fi
+
+	if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+	then
 	mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png
+	cp ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png $COMOUTplots/$varb
+        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.fhrmean.firewx.png ]
+	then
+	echo "NO PLOT FOR",$var,$plottyp
+	fi
 
 	export plottyp=valid_hour
 	export datetyp=INIT
-	sh $USHevs/${COMPONENT}/py_plotting.config_pbl
 
+	if [ ! -e $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png ]
+	then
+	sh $USHevs/${COMPONENT}/py_plotting.config_pbl
+        else
+	echo "RESTART - plot exists; copying over to plot directory"
+	cp $COMOUTplots/$varb/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png $PLOTDIR
+	fi
+
+	if [ -e ${PLOTDIR}/sfc_upper/*/evs*png ]
+	then
 	mv ${PLOTDIR}/sfc_upper/*/evs*png ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png
+	cp ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png $COMOUTplots/$varb
+        elif [ ! -e ${PLOTDIR}/evs.${MODELNAME}.bcrmse_me.${smvar}_${smlev}.last31days.vhrmean.firewx.png ]
+	then
+	echo "NO PLOT FOR",$var,$plottyp
+	fi
 
 done
